# [PERF] TSDB: Grow postings by doubling