@bboreham 2024-09-02
I will try to get round to benchmarks soon.