@bboreham 2024-09-23
I spent some time wrestling with `BenchmarkCreateSeries`, because as it stands it does not do the same work for each iteration (adding series slows down as the number of existing series grows).  But across several existing benchmarks with `Create` in the name we see a reasonable picture: slightly faster, and less memory allocation.